+++ hal/display/pico/pico_ili9341_pivro_display_hal.c
@@ -8,7 +8,7 @@ void pivro_display_initialize()
 }
 
 void pivro_display_draw_buffer(
-    uint8_t *buffer,
+    uint16_t* buffer,
     size_t size)
 {
     ili9341_display_draw_buffer(buffer, size);
